requirements.txt
aiofiles>=24.1.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
dnspython>=2.6.0
ffmpeg-python==0.2.0
gitpython>=3.1.40
hachoir>=3.2.1
heroku3==5.2.1
httpx>=0.25.0
motor>=3.3.2
numpy>=1.24.0
opencv-python>=4.8.0
pillow>=11.0.0
psutil>=6.0.0
pyrogram>=2.0.106
py-tgcalls>=2.2.11
pykeyboard==0.1.5
python-dotenv>=1.0.0
pyyaml>=6.0.0
requests>=2.31.0
speedtest-cli>=2.1.3
spotipy>=2.22.1
pymongo>=4.6.0
tgcrypto>=1.2.5;sys_platform != "win32"
unidecode>=1.3.0
uvloop>=0.19.0;sys_platform != "win32"
yt-dlp>=2025.2.10
youtube-search>=2.1.1
pytube>=15.0.0
colorama>=0.4.6
tenacity>=8.2.3
